Making the whole thing smaller
Makes the funnel less deep, and 1/2" smaller. This allows for the 3"
filter to come up on the sides and hang outside, ensuring a seal around
the whole thing. The notches for the rubber band didn't need to be so
big, either, and this fixes the grill so it's along the edges as well.